separate logistic and linear
