+++ src/leaspy/models/__init__.py
@@ -2,13 +2,11 @@ from .base import BaseModel, ModelInterface
 from .constant import ConstantModel
 from .factory import ModelName, model_factory
 from .joint import JointModel
+from .linear import LinearModel
 from .lme import LMEModel
+from .logistic import LogisticModel
 from .mcmc_saem_compatible import McmcSaemCompatibleModel
-from .riemanian_manifold import (
-    LinearModel,
-    LogisticModel,
-    RiemanianManifoldModel,
-)
+from .riemanian_manifold import RiemanianManifoldModel
 from .settings import ModelSettings
 from .shared_speed_logistic import SharedSpeedLogisticModel
 from .stateful import StatefulModel
