+++ src/leaspy/models/joint.py
@@ -21,8 +21,8 @@ from leaspy.variables.specs import (
 )
 from leaspy.variables.state import State
 
+from .logistic import LogisticModel
 from .obs_models import observation_model_factory
-from .riemanian_manifold import LogisticModel
 
 __all__ = ["JointModel"]
